Fix for portable mode with "shred settings and quit"
git-svn-id: http://svn.code.sf.net/p/bleachbit/code/trunk@1743 173665bb-4496-40ae-b72c-49697fe1074e

+++ bleachbit/GUI.py
@@ -509,15 +509,16 @@ class GUI:
             self.append_text(err, 'error')
             return
 
-        try:
-            os.rmdir(options_dir)
-        except:
-            print traceback.print_exc()
-            err = str(sys.exc_info()[1])
-            self.append_text(err, 'error')
-            while gtk.events_pending():
-                gtk.main_iteration()
-            time.sleep(3)
+        if not portable_mode:
+            try:
+                os.rmdir(options_dir)
+            except:
+                print traceback.print_exc()
+                err = str(sys.exc_info()[1])
+                self.append_text(err, 'error')
+                while gtk.events_pending():
+                    gtk.main_iteration()
+                time.sleep(3)
 
         gtk.main_quit()
